# src/main/resources/openapi/openapi.yml
openapi: 3.0.0
info:
  title: API for dream case crud
  version: 1.0.0
paths:
  /api/v1/cases/{caseID}:
    get:
      summary: Retrieve case by id
      operationId: readCase
      tags:
        - cases
      parameters:
        - $ref: './commons/commons.yaml/#/components/parameters/caseID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: './commons/commons.yaml#/components/schemas/Case'
        default:
          description: error
          content:
            application/json:
              schema:
                $ref: "./commons/commons.yaml#/components/schemas/Error"
    put:
      summary: update case
      operationId: updateCase
      tags:
        - cases
      parameters:
        - $ref: './commons/commons.yaml/#/components/parameters/caseID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./commons/commons.yaml/#/components/schemas/Case"
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: './commons/commons.yaml#/components/schemas/Case'
        default:
          description: error
          content:
            application/json:
              schema:
                $ref: "./commons/commons.yaml#/components/schemas/Error"
    delete:
      summary: Delete case by id
      operationId: deleteCase
      tags:
        - cases
      parameters:
        - $ref: './commons/commons.yaml/#/components/parameters/caseID'
      responses:
        '200':
          description: Successful operation
        default:
          description: error
          content:
            application/json:
              schema:
                $ref: "./commons/commons.yaml#/components/schemas/Error"

  /api/v1/cases:
    post:
      summary: create case
      operationId: createCase
      tags:
        - cases
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./commons/commons.yaml/#/components/schemas/Case"
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: './commons/commons.yaml#/components/schemas/Case'
        default:
          description: error
          content:
            application/json:
              schema:
                $ref: "./commons/commons.yaml#/components/schemas/Error"